LayoutTests/imported/w3c/web-platform-tests/html/webappapis/scripting/events/inline-event-handler-ordering-expected.txt

CONSOLE MESSAGE: SyntaxError: Unexpected token '}'
CONSOLE MESSAGE: SyntaxError: Unexpected token '}'

PASS Inline event handlers retain their ordering when invalid and force-compiled
PASS Inline event handlers retain their ordering when invalid and force-compiled via dispatch
PASS Inline event handlers retain their ordering when invalid and lazy-compiled
